requirements.txt
pymongo==4.13.1
requests==2.32.4
sentence-transformers==4.1.0
python-dotenv==1.1.0